@@ -3,20 +3,14 @@ package main
 import (
 	"flag"
 	"fmt"
-	"io/ioutil"
+	"os"
 
-	"github.com/jrmanes/torch/config"
-	handlers "github.com/jrmanes/torch/pkg/http"
-
-	"github.com/jrmanes/torch/pkg/k8s"
 	log "github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v2"
-)
 
-// Configuration variables
-var (
-	// cfg stores the mutual peers configuration.
-	cfg config.MutualPeersConfig
+	"github.com/celestiaorg/torch/config"
+	handlers "github.com/celestiaorg/torch/pkg/http"
+	"github.com/celestiaorg/torch/pkg/k8s"
 )
 
 // ParseFlags parses the command-line flags and reads the configuration file.
@@ -28,10 +22,12 @@ func ParseFlags() config.MutualPeersConfig {
 	flag.Parse()
 
 	// Read the configuration file
-	file, err := ioutil.ReadFile(*configFile)
+	file, err := os.ReadFile(*configFile)
 	if err != nil {
 		log.Error("Config file doesn't exist...", err)
 	}
+
+	cfg := config.MutualPeersConfig{}
 
 	// Unmarshal the YAML into a struct
 	err = yaml.Unmarshal(file, &cfg)
@@ -59,7 +55,7 @@ func main() {
 	PrintName()
 	// Parse the command-line flags and read the configuration file
 	log.Info("Running on namespace: ", k8s.GetCurrentNamespace())
-	cfg = ParseFlags()
+	cfg := ParseFlags()
 
 	handlers.Run(cfg)
 }
