@@ -1,12 +1,11 @@
-package k8s
+package metrics
 
 import (
-	"github.com/jrmanes/torch/pkg/metrics"
 	log "github.com/sirupsen/logrus"
 )
 
-// Declare a slice to hold multiple MultiAddrs metrics.
-var multiAddresses []metrics.MultiAddrs
+// Declare a slice to hold multiple Multi Addresses metrics.
+var multiAddresses []MultiAddrs
 
 // MultiAddrExists checks if a given MultiAddr already exists in the multiAddresses slice.
 // It returns true if the MultiAddr already exists, and false otherwise.
@@ -20,11 +19,11 @@ func MultiAddrExists(multiAddr string) bool {
 	return false
 }
 
-// RegisterMetric adds a new MultiAddrs metric to the multiAddresses slice.
+// RegisterMetric adds a new Multi Addresses metric to the multiAddresses slice.
 // Before adding, it checks if the MultiAddr already exists in the slice using MultiAddrExists function.
 // If the MultiAddr already exists, it logs a message and skips the addition.
-// Otherwise, it appends the new MultiAddrs to the slice and registers the updated metrics.
-func RegisterMetric(m metrics.MultiAddrs) {
+// Otherwise, it appends the new Multi Addresses to the slice and registers the updated metrics.
+func RegisterMetric(m MultiAddrs) {
 	// Check if the MultiAddr already exists in the array
 	if MultiAddrExists(m.MultiAddr) {
 		log.Info("MultiAddr already exists in the metrics array: ", m.NodeName, " ", m.MultiAddr)
@@ -35,7 +34,7 @@ func RegisterMetric(m metrics.MultiAddrs) {
 	multiAddresses = append(multiAddresses, m)
 
 	// Register the metric
-	err := metrics.WithMetricsMultiAddress(multiAddresses)
+	err := WithMetricsMultiAddress(multiAddresses)
 	if err != nil {
 		log.Printf("Failed to update metrics: %v", err)
 	}
